WIP: Updating users controller.
This commit works on updating the users controller for form
validation. All the form validation is working correctly. It makes
sure all the inputs have been put in properly. If there is any
issues an error message displayed under the field as well as making
the field red. Both forms are working properly. I tested them.
Database inserts and lookup still have to be implemented.