Checking if email already exists.
This commit updates the users controller to look up the email
in the database to see if it already exists if a user registering
provides an email. If it already exists it retturns an error
saying that email is already taken. This commit also updates the
constructor to pull in the model onto the controller. This commit
also creates the user model which implements a function to find
a user by email. This function uses the methods on the Database
lib class to find user record(s).